Added countdown clock/ Final commit
Added The variables set the value to 30 min and 0 seconds. Added the variable that makes The time left on the clock will by multiply by 60 and added to the seconds left. The clock will count down by 1000 milliseconds. The leading zero will make sure that the leading digit will be a 0 unless sec/minutes is above ten . Changes the minutes/seconds string to the proper value. Added var where When the clock hits zero and if form is not done, the message "Order Expired" will show. var where clock from rerunning when clock hits zero.  Ready for review.